Support for DELETE method and non-200 responses
The Delete method adds support for DELETE requests.  Pretty basic.

There is a new option that can be set when issuing API calls.  Some API
endpoints return non-200 results when they succeed; for example, API
calls which do not return a response are allowed to respond with a 204
result code.  To support this without thinking it's an error, the
OkCodes field may now be initialized with a slice of valid response
codes.  If none is provided or specified, the default behavior is to
assume only a 200 response to be valid.