Add raw response JSON feedback path.
While debugging a problem with Gorax, I needed to see raw JSON, but
didn't want to interject printf-statements everywhere, or include
customized logic to determine _when_ to print or not print.  The proper
solution was to just get a copy of the raw JSON results back, but only
when I needed them.

Thus, the ResponseJson option was introduced to provide that
information on an on-demand basis.