Retrofit request() to return HTTP.Response
Direct access to arbitrary headers is a significant concern for many
OpenStack APIs.  The current API does not offer this access.
Retrofitting request() to grant access to the raw Response is the first
step towards achieving this goal through the public API.

It should be noted that the public API does not yet expose this
functionality.  This will be addressed in a separate commit.